refactor(engine): move GraphExt to own file, fmt
