+++ crates/nvisy-engine/src/operation/import_file.rs
@@ -3,7 +3,7 @@
 //! Runs at **phase 0**. Decodes raw content into a
 //! [`DocumentEnvelope`], optionally applying decompression and
 //! decryption.
-//! 
+//!
 //! The import pipeline applies optional pre-processing steps in order:
 //!
 //! 1. **Decompression** — decompress raw bytes (if format specified)
