+++ crates/nvisy-engine/src/pipeline/run.rs
@@ -22,8 +22,8 @@ use super::config::RuntimeConfig;
 use super::default::{EngineInput, EngineOutput};
 use super::orchestrator::{Orchestrator, RunContext};
 use super::plan::{self, ExecutionPlan};
-use super::runs::state::{RunRecord, RunState};
 use super::runs::RunStatus;
+use super::runs::state::{RunRecord, RunState};
 use crate::operation::encryption::SharedKeyProvider;
 use crate::operation::envelope::SharedData;
 use crate::registry::Registry;
